+++ src/utils/useTableDimension.ts
@@ -10,8 +10,8 @@ import isNumberOrTrue from './isNumberOrTrue';
 import { RowDataType, RowKeyType, ElementOffset } from '../@types/common';
 
 interface TableDimensionProps {
-  data?: RowDataType[];
-  rowHeight?: number | ((rowData: RowDataType) => number);
+  data?: readonly RowDataType[];
+  rowHeight: number | ((rowData: RowDataType) => number);
   height: number;
   minHeight: number;
   tableRef?: React.RefObject<HTMLDivElement>;
@@ -26,10 +26,12 @@ interface TableDimensionProps {
   fillHeight?: boolean;
   children?: React.ReactNode;
   expandedRowKeys?: RowKeyType[];
+  showHeader?: boolean;
   onTableScroll?: (coord: { x?: number; y?: number }) => void;
-  onTableContentHeightChange?: (prevHeight: number) => void;
-  onTableContentWidthChange?: (prevWidth: number) => void;
-  onTableWidthChange?: (prevWidth: number) => void;
+  onTableResizeChange?: (
+    prevSize: number,
+    event: 'bodyHeightChanged' | 'bodyWidthChanged' | 'widthChanged' | 'heightChanged'
+  ) => void;
 }
 
 /**
@@ -56,9 +58,8 @@ const useTableDimension = (props: TableDimensionProps) => {
     fillHeight,
     children,
     expandedRowKeys,
-    onTableWidthChange,
-    onTableContentWidthChange,
-    onTableContentHeightChange,
+    showHeader,
+    onTableResizeChange,
     onTableScroll
   } = props;
 
@@ -69,25 +70,34 @@ const useTableDimension = (props: TableDimensionProps) => {
   const scrollX = useRef(0);
   const minScrollX = useRef(0);
   const tableWidth = useRef(widthProp || 0);
+  const tableHeight = useRef(heightProp || 0);
   const columnCount = useRef(0);
   const resizeObserver = useRef<ResizeObserver>();
   const containerResizeObserver = useRef<ResizeObserver>();
   const headerOffset = useRef<ElementOffset | null>(null);
   const tableOffset = useRef<ElementOffset | null>(null);
 
-  const [tableHeight, setTableHeight] = useState(heightProp || 0);
   const [emptySpaceBelow, setEmptySpaceBelow] = useState(0);
-  const tableHeightWithoutFooter = tableHeight - footerHeight;
+  const tableHeightWithoutFooter = tableHeight.current - footerHeight;
+  const getRowHeight = useCallback(
+    (rowData = {}) => {
+      return typeof rowHeight === 'function' ? rowHeight(rowData) : rowHeight;
+    },
+    [rowHeight]
+  );
 
   const calculateTableContextHeight = useCallback(() => {
     const prevContentHeight = contentHeight.current;
-    const table = tableRef?.current;
+    const table = tableRef?.current as HTMLDivElement;
     const rows = table?.querySelectorAll(`.${prefix?.('row')}`) || [];
+    const virtualized = table?.querySelectorAll('.virtualized')?.length > 0;
 
     const nextContentHeight = rows.length
-      ? (Array.from(rows).map((row: Element) => getHeight(row) || rowHeight) as number[]).reduce(
-          (x: number, y: number) => x + y
-        )
+      ? (
+          Array.from(rows).map(
+            (row: Element, index: number) => getHeight(row) || getRowHeight(data?.[index])
+          ) as number[]
+        ).reduce((x: number, y: number) => x + y)
       : 0;
 
     // After setting the affixHeader property, the height of the two headers should be subtracted.
@@ -95,14 +105,26 @@ const useTableDimension = (props: TableDimensionProps) => {
       nextContentHeight - (affixHeader ? headerHeight * 2 : headerHeight)
     );
 
+
+
+    // Whether to show the horizontal scroll bar
+    const hasHorizontalScrollbar = contentWidth.current > tableWidth.current;
+
+    // The height of the table content area should be added to the height occupied by the horizontal scroll bar when autoHeight is set.
+    if (autoHeight && hasHorizontalScrollbar) {
+      contentHeight.current += SCROLLBAR_WIDTH;
+    }
+
+    const height = fillHeight ? tableHeight.current : heightProp;
+    const tableBodyHeight = showHeader ? height - headerHeight : height;
+
     if (!autoHeight) {
       /**
        *  The purpose of subtracting SCROLLBAR_WIDTH is to keep the scroll bar from blocking the content part.
        *  But it will only be calculated when there is a horizontal scroll bar (contentWidth > tableWidth).
        */
       minScrollY.current =
-        -(nextContentHeight - tableHeightWithoutFooter) -
-        (contentWidth.current > tableWidth.current ? SCROLLBAR_WIDTH : 0);
+        -(nextContentHeight - tableHeightWithoutFooter) - (hasHorizontalScrollbar ? SCROLLBAR_WIDTH : 0);
     }
 
     // If the height of the content area is less than the height of the table, the vertical scroll bar is reset.
@@ -121,10 +143,24 @@ const useTableDimension = (props: TableDimensionProps) => {
       nextContentHeight + SCROLLBAR_WIDTH
     ) {
       onTableScroll?.({ y: scrollY.current });
+    const currentScrollTop = Math.abs(scrollY.current);
+
+    // When Table is set to virtualized, the logic will be entered every time the wheel event is
+    // triggered to avoid resetting the scroll bar after scrolling to the bottom, so add the SCROLLBAR_WIDTH value.
+    const maxScrollTop = nextContentHeight + SCROLLBAR_WIDTH - tableBodyHeight;
+
+    // If the top value of the current scroll is greater than the scrollable range,
+    // keep the vertical scroll bar at the bottom.
+    if (maxScrollTop > 0 && currentScrollTop > maxScrollTop) {
+      if (virtualized) {
+        onTableScroll?.({ y: (data?.length || 0) * getRowHeight() - tableBodyHeight });
+      } else {
+        onTableScroll?.({ y: maxScrollTop });
+      }
     }
 
     if (prevContentHeight !== contentHeight.current) {
-      onTableContentHeightChange?.(prevContentHeight);
+      onTableResizeChange?.(prevContentHeight, 'bodyHeightChanged');
     }
   }, [
     tableRef,
@@ -132,10 +168,13 @@ const useTableDimension = (props: TableDimensionProps) => {
     affixHeader,
     headerHeight,
     autoHeight,
-    tableHeight,
-    rowHeight,
+    fillHeight,
+    heightProp,
+    showHeader,
+    getRowHeight,
+    data,
     onTableScroll,
-    onTableContentHeightChange
+    onTableResizeChange
   ]);
 
   const setOffsetByAffix = useCallback(() => {
@@ -161,7 +200,8 @@ const useTableDimension = (props: TableDimensionProps) => {
 
     // The value of SCROLLBAR_WIDTH is subtracted so that the scroll bar does not block the content part.
     // There is no vertical scroll bar after autoHeight.
-    minScrollX.current = -(nextContentWidth - tableWidth.current);
+    minScrollX.current =
+      -(nextContentWidth - tableWidth.current) - (autoHeight ? 0 : SCROLLBAR_WIDTH);
 
     /**
      * If the width of the content area and the number of columns change,
@@ -174,9 +214,9 @@ const useTableDimension = (props: TableDimensionProps) => {
       prevColumnCount > 0 &&
       prevColumnCount !== columnCount.current
     ) {
-      onTableContentWidthChange?.(prevWidth);
+      onTableResizeChange?.(prevWidth, 'bodyWidthChanged');
     }
-  }, [autoHeight, onTableContentWidthChange, prefix, tableRef]);
+  }, [autoHeight, onTableResizeChange, prefix, tableRef]);
 
   const calculateTableWidth = useCallback(
     (nextWidth?: number) => {
@@ -188,57 +228,72 @@ const useTableDimension = (props: TableDimensionProps) => {
 
       if (prevWidth && prevWidth !== tableWidth.current) {
         scrollX.current = 0;
-        onTableWidthChange?.(prevWidth);
+        onTableResizeChange?.(prevWidth, 'widthChanged');
       }
 
       setOffsetByAffix();
     },
-    [onTableWidthChange, setOffsetByAffix, tableRef]
+    [onTableResizeChange, setOffsetByAffix, tableRef]
+  );
+
+  const calculateTableHeight = useCallback(
+    (nextHeight?: number) => {
+      const prevHeight = tableHeight.current;
+
+      if (nextHeight) {
+        tableHeight.current = nextHeight;
+      } else if (tableRef?.current) {
+        tableHeight.current = getHeight(tableRef.current.parentNode as Element);
+      }
+
+      if (prevHeight && prevHeight !== tableHeight.current) {
+        onTableResizeChange?.(prevHeight, 'heightChanged');
+      }
+    },
+    [onTableResizeChange, tableRef]
   );
 
   useMount(() => {
     calculateTableContextHeight();
     calculateTableContentWidth();
     calculateTableWidth();
+    calculateTableHeight();
     setOffsetByAffix();
 
-    // When fillHeight is set, a resize listener is added to the table container.
-    // And get the height of the container as the height of the table.
-    if (fillHeight && tableRef?.current) {
-      setTableHeight(getHeight(tableRef?.current.parentNode?.parentNode as Element));
-      containerResizeObserver.current = new ResizeObserver(entries => {
-        setTableHeight(entries[0].contentRect.height);
-        calculateTableWidth(entries[0].contentRect.width);
-      });
-      containerResizeObserver.current.observe(tableRef?.current?.parentNode?.parentNode as Element);
-    } else {
-      resizeObserver.current = new ResizeObserver(entries => {
-        calculateTableWidth(entries[0].contentRect.width);
-      });
-      resizeObserver.current.observe(tableRef?.current as Element);
-    }
+    containerResizeObserver.current = new ResizeObserver(entries => {
+      calculateTableHeight(entries[0].contentRect.height);
+    });
+    containerResizeObserver.current.observe(tableRef?.current?.parentNode as Element);
+
+    resizeObserver.current = new ResizeObserver(entries => {
+      calculateTableWidth(entries[0].contentRect.width);
+    });
+    resizeObserver.current.observe(tableRef?.current as Element);
+
+    return () => {
+      resizeObserver.current?.disconnect();
+      containerResizeObserver.current?.disconnect();
+    };
   });
 
   useUpdateLayoutEffect(() => {
-    calculateTableWidth();
+    calculateTableHeight();
     calculateTableContextHeight();
+  }, [fillHeight]);
+
+  useUpdateLayoutEffect(() => {
+    calculateTableWidth();
     calculateTableContentWidth();
+    calculateTableContextHeight();
   }, [
     data,
     heightProp,
-    contentHeight,
+    contentHeight.current,
     expandedRowKeys,
     children,
     calculateTableContextHeight,
     calculateTableContentWidth
   ]);
-
-  useEffect(() => {
-    return () => {
-      resizeObserver.current?.disconnect();
-      containerResizeObserver.current?.disconnect();
-    };
-  }, []);
 
   const setScrollY = useCallback((value: number) => {
     scrollY.current = value;
@@ -250,7 +305,7 @@ const useTableDimension = (props: TableDimensionProps) => {
 
   const getTableHeight = () => {
     if (fillHeight) {
-      return tableHeight;
+      return tableHeight.current;
     }
 
     if (data?.length === 0 && autoHeight) {